+++ src/tests/example/test_module.py
@@ -1,4 +1,0 @@
-from dependency.core import Module, module
-
-def test_module():
-    pass