+++ tests/core/test_exceptions.py
@@ -2,7 +2,7 @@ import pytest
 from dependency_injector import containers
 from dependency.core.agrupation import Plugin, PluginMeta, module
 from dependency.core.declaration import Component, component, Product, product, instance
-from dependency.core.injection import InjectionLoader, Container
+from dependency.core.resolution import Container, InjectionResolver
 from dependency.core.exceptions import DeclarationError, ResolutionError
 
 @module()
@@ -27,7 +27,7 @@ class TComponent2(Component):
     pass
 
 @product(
-    imports=[TComponent1]
+    imports=[TComponent1],
 )
 class TProduct1(Product):
     pass
@@ -35,7 +35,7 @@ class TProduct1(Product):
 @instance(
     component=TComponent1,
     imports=[TComponent2],
-    products=[TProduct1]
+    products=[TProduct1],
 )
 class TInstance1(TInterface):
     pass
@@ -52,13 +52,13 @@ def test_exceptions():
     with pytest.raises(ResolutionError):
         TPlugin.resolve_providers(container) # type: ignore
 
-    container = Container.from_json("example/config.json")
+    container = Container()
     providers = TPlugin.resolve_providers(container) # type: ignore
-    loader = InjectionLoader(container, providers)
+    loader = InjectionResolver(container, providers)
 
+    with pytest.raises(DeclarationError):
+        print(TComponent1.provider())
     with pytest.raises(DeclarationError):
         print(TComponent1.provide())
     with pytest.raises(ResolutionError):
-        loader.resolve_providers()
-    with pytest.raises(ResolutionError):
-        loader.resolve_products()
+        loader.resolve_injectables()